backend/requirements.txt
fastapi
uvicorn
motor
google-generativeai
python-dotenv
pydantic
pydantic-settings